requirements.txt
discord.py>=0.15.0,<0.17.0
basehash>=2.0.0,<3.4.0